moved SortableDomRect to utils file
