+++ types/sortable-utils.d.ts
@@ -72,3 +72,15 @@ export interface SortableUtils {
 	 */
 	detectDirection(element: HTMLElement): 'vertical' | 'horizontal';
 }
+
+
+export interface SortableDOMRect {
+	bottom: number;
+	height: number;
+	left: number;
+	right: number;
+	top: number;
+	width: number;
+	x: number;
+	y: number;
+}
